Moved DefaultAbsorptionTimes data to LoopSettings
