+++ Loop/Extensions/NSUserDefaults.swift
@@ -13,14 +13,13 @@ import HealthKit
 import CarbKit
 
 extension UserDefaults {
-    
+
     private enum Key: String {
         case basalRateSchedule = "com.loudnate.Naterade.BasalRateSchedule"
         case batteryChemistry = "com.loopkit.Loop.BatteryChemistry"
         case cgmSettings = "com.loopkit.Loop.cgmSettings"
         case carbRatioSchedule = "com.loudnate.Naterade.CarbRatioSchedule"
         case connectedPeripheralIDs = "com.loudnate.Naterade.ConnectedPeripheralIDs"
-        case defaultAbsorptionTimes = "com.loudnate.Naterade.DefaultAbsorptionTimes"
         case loopSettings = "com.loopkit.Loop.loopSettings"
         case insulinActionDuration = "com.loudnate.Naterade.InsulinActionDuration"
         case insulinCounteractionEffects = "com.loopkit.Loop.insulinCounteractionEffects"
@@ -31,7 +30,7 @@ extension UserDefaults {
         case pumpRegion = "com.loopkit.Loop.PumpRegion"
         case pumpTimeZone = "com.loudnate.Naterade.PumpTimeZone"
     }
-    
+
     var basalRateSchedule: BasalRateSchedule? {
         get {
             if let rawValue = dictionary(forKey: Key.basalRateSchedule.rawValue) {
@@ -44,7 +43,7 @@ extension UserDefaults {
             set(newValue?.rawValue, forKey: Key.basalRateSchedule.rawValue)
         }
     }
-    
+
     var carbRatioSchedule: CarbRatioSchedule? {
         get {
             if let rawValue = dictionary(forKey: Key.carbRatioSchedule.rawValue) {
@@ -57,7 +56,7 @@ extension UserDefaults {
             set(newValue?.rawValue, forKey: Key.carbRatioSchedule.rawValue)
         }
     }
-    
+
     var cgm: CGM? {
         get {
             if let rawValue = dictionary(forKey: Key.cgmSettings.rawValue) {
@@ -70,22 +69,22 @@ extension UserDefaults {
                     removeObject(forKey: "com.loopkit.Loop.FetchEnliteDataEnabled")
                     removeObject(forKey: "com.loudnate.Naterade.TransmitterID")
                 }
-                
+
                 if bool(forKey: "com.loudnate.Loop.G4ReceiverEnabled") {
                     self.cgm = .g4
                     return .g4
                 }
-                
+
                 if bool(forKey: "com.loopkit.Loop.FetchEnliteDataEnabled") {
                     self.cgm = .enlite
                     return .enlite
                 }
-                
+
                 if let transmitterID = string(forKey: "com.loudnate.Naterade.TransmitterID"), transmitterID.characters.count == 6 {
                     self.cgm = .g5(transmitterID: transmitterID)
                     return .g5(transmitterID: transmitterID)
                 }
-                
+
                 return nil
             }
         }
@@ -93,7 +92,7 @@ extension UserDefaults {
             set(newValue?.rawValue, forKey: Key.cgmSettings.rawValue)
         }
     }
-    
+
     var connectedPeripheralIDs: [String] {
         get {
             return array(forKey: Key.connectedPeripheralIDs.rawValue) as? [String] ?? []
@@ -102,32 +101,7 @@ extension UserDefaults {
             set(newValue, forKey: Key.connectedPeripheralIDs.rawValue)
         }
     }
-    
-    var defaultAbsorptionTimes: CarbStore.DefaultAbsorptionTimes? {
-        get {
-            if let defaultAbsorptionTimesDict = dictionary(forKey: Key.defaultAbsorptionTimes.rawValue) as? [String : TimeInterval] {
-                return (fast: defaultAbsorptionTimesDict["fast"],
-                        medium: defaultAbsorptionTimesDict["medium"],
-                        slow: defaultAbsorptionTimesDict["slow"]) as? CarbStore.DefaultAbsorptionTimes
-            } else {
-                return nil
-            }
-        }
-        set {
-            // tuples cannot be stored directly in UserDefaults
-            if let newTimes = newValue {
-                let defaultAbsorptionTimesDict: [String : TimeInterval] = [
-                    "fast" : newTimes.fast,
-                    "medium" : newTimes.medium,
-                    "slow" : newTimes.slow
-                ]
-                set(defaultAbsorptionTimesDict, forKey: Key.defaultAbsorptionTimes.rawValue)
-            } else {
-                removeObject(forKey: Key.defaultAbsorptionTimes.rawValue)
-            }
-        }
-    }
-    
+
     var loopSettings: LoopSettings? {
         get {
             if let rawValue = dictionary(forKey: Key.loopSettings.rawValue) {
@@ -142,33 +116,40 @@ extension UserDefaults {
                     removeObject(forKey: "com.loopkit.Loop.MinimumBGGuard")
                     removeObject(forKey: "com.loudnate.Loop.RetrospectiveCorrectionEnabled")
                 }
-                
+
+                let defaultAbsorptionTimes = (
+                    fast: TimeInterval(hours: 2),
+                    medium: TimeInterval(hours: 3),
+                    slow: TimeInterval(hours: 4)
+                )
+
                 let glucoseTargetRangeSchedule: GlucoseRangeSchedule?
                 if let rawValue = dictionary(forKey: "com.loudnate.Naterade.GlucoseTargetRangeSchedule") {
                     glucoseTargetRangeSchedule = GlucoseRangeSchedule(rawValue: rawValue)
                 } else {
                     glucoseTargetRangeSchedule = nil
                 }
-                
+
                 let minimumBGGuard: GlucoseThreshold?
                 if let rawValue = dictionary(forKey: "com.loopkit.Loop.MinimumBGGuard") {
                     minimumBGGuard = GlucoseThreshold(rawValue: rawValue)
                 } else {
                     minimumBGGuard = nil
                 }
-                
+
                 var maximumBasalRatePerHour: Double? = double(forKey: "com.loudnate.Naterade.MaximumBasalRatePerHour")
                 if maximumBasalRatePerHour! <= 0 {
                     maximumBasalRatePerHour = nil
                 }
-                
+
                 var maximumBolus: Double? = double(forKey: "com.loudnate.Naterade.MaximumBolus")
                 if maximumBolus! <= 0 {
                     maximumBolus = nil
                 }
-                
+
                 let settings = LoopSettings(
                     dosingEnabled: bool(forKey: "com.loudnate.Naterade.DosingEnabled"),
+                    defaultAbsorptionTimes: defaultAbsorptionTimes,
                     glucoseTargetRangeSchedule: glucoseTargetRangeSchedule,
                     maximumBasalRatePerHour: maximumBasalRatePerHour,
                     maximumBolus: maximumBolus,
@@ -176,7 +157,7 @@ extension UserDefaults {
                     retrospectiveCorrectionEnabled: bool(forKey: "com.loudnate.Loop.RetrospectiveCorrectionEnabled")
                 )
                 self.loopSettings = settings
-                
+
                 return settings
             }
         }
@@ -184,7 +165,7 @@ extension UserDefaults {
             set(newValue?.rawValue, forKey: Key.loopSettings.rawValue)
         }
     }
-    
+
     var insulinActionDuration: TimeInterval? {
         get {
             let value = double(forKey: Key.insulinActionDuration.rawValue)
@@ -199,7 +180,7 @@ extension UserDefaults {
             }
         }
     }
-    
+
     var insulinSensitivitySchedule: InsulinSensitivitySchedule? {
         get {
             if let rawValue = dictionary(forKey: Key.insulinSensitivitySchedule.rawValue) {
@@ -212,7 +193,7 @@ extension UserDefaults {
             set(newValue?.rawValue, forKey: Key.insulinSensitivitySchedule.rawValue)
         }
     }
-    
+
     var preferredInsulinDataSource: InsulinDataSource? {
         get {
             return InsulinDataSource(rawValue: integer(forKey: Key.preferredInsulinDataSource.rawValue))
@@ -225,7 +206,7 @@ extension UserDefaults {
             }
         }
     }
-    
+
     var pumpID: String? {
         get {
             return string(forKey: Key.pumpID.rawValue)
@@ -234,7 +215,7 @@ extension UserDefaults {
             set(newValue, forKey: Key.pumpID.rawValue)
         }
     }
-    
+
     var pumpModelNumber: String? {
         get {
             return string(forKey: Key.pumpModelNumber.rawValue)
@@ -243,7 +224,7 @@ extension UserDefaults {
             set(newValue, forKey: Key.pumpModelNumber.rawValue)
         }
     }
-    
+
     var pumpRegion: PumpRegion? {
         get {
             // Defaults to 0 / northAmerica
@@ -253,7 +234,7 @@ extension UserDefaults {
             set(newValue?.rawValue, forKey: Key.pumpRegion.rawValue)
         }
     }
-    
+
     var pumpTimeZone: TimeZone? {
         get {
             if let offset = object(forKey: Key.pumpTimeZone.rawValue) as? NSNumber {
@@ -269,7 +250,7 @@ extension UserDefaults {
             }
         }
     }
-    
+
     var batteryChemistry: BatteryChemistryType? {
         get {
             return BatteryChemistryType(rawValue: integer(forKey: Key.batteryChemistry.rawValue))
@@ -282,5 +263,5 @@ extension UserDefaults {
             }
         }
     }
-    
+
 }
